refactor(geo-map): slim API and encapsulate Leaflet engine
