+++ app/docs/geo-map/content.mdx
@@ -230,33 +230,13 @@ Supported token keys:
       description: "Tailwind classes for the root wrapper",
       type: "string",
     },
-    mapClassName: {
-      description: "Tailwind classes for the map frame wrapper",
-      type: "string",
-    },
     style: {
       description:
         "Inline style overrides for GeoMap CSS custom properties (for popup/tooltip shell theming)",
-      type: "GeoMapStyle",
-    },
-    overlayClassName: {
-      description: "Tailwind classes for the title/description overlay panel",
-      type: "string",
+      type: "CSSProperties",
     },
     popupClassName: {
       description: "Tailwind classes applied to the Leaflet popup shell",
-      type: "string",
-    },
-    popupContentClassName: {
-      description: "Tailwind classes for popup content layout container",
-      type: "string",
-    },
-    popupTitleClassName: {
-      description: "Tailwind classes for popup title text",
-      type: "string",
-    },
-    popupDescriptionClassName: {
-      description: "Tailwind classes for popup description text",
       type: "string",
     },
     tooltipClassName: {
