+++ components/tool-ui/geo-map/README.md
@@ -6,9 +6,9 @@ Implementation for the "geo-map" Tool UI surface.
 
 - public exports: components/tool-ui/geo-map/index.tsx
 - serializable schema + parse helpers: components/tool-ui/geo-map/schema.ts
-- main runtime orchestrator: components/tool-ui/geo-map/geo-map.tsx
+- public facade component: components/tool-ui/geo-map/geo-map.tsx
+- internal Leaflet engine: components/tool-ui/geo-map/geo-map-engine.tsx
 - colocated Leaflet shell theme styles: components/tool-ui/geo-map/geo-map-theme.module.css
-- viewport sync + controllers: components/tool-ui/geo-map/geo-map-viewport.tsx
 - icon construction helpers: components/tool-ui/geo-map/geo-map-icons.ts
 - popup/tooltip overlay renderer: components/tool-ui/geo-map/geo-map-overlays.tsx
