+++ components/tool-ui/geo-map/geo-map-overlays.tsx
@@ -7,35 +7,19 @@ import { Popup, Tooltip, cn } from "./_adapter";
 function GeoMapPopupContent({
   label,
   description,
-  className,
-  titleClassName,
-  descriptionClassName,
 }: {
   label?: string;
   description?: string;
-  className?: string;
-  titleClassName?: string;
-  descriptionClassName?: string;
 }) {
   return (
-    <div className={cn("flex flex-col gap-0.5", className)}>
+    <div className="flex flex-col gap-0.5">
       {label && (
-        <p
-          className={cn(
-            "block text-sm leading-tight font-semibold tracking-tight text-foreground",
-            titleClassName,
-          )}
-        >
+        <p className="block text-sm leading-tight font-semibold tracking-tight text-foreground">
           {label}
         </p>
       )}
       {description && (
-        <p
-          className={cn(
-            "block text-xs leading-relaxed text-muted-foreground",
-            descriptionClassName,
-          )}
-        >
+        <p className="block text-xs leading-relaxed text-muted-foreground">
           {description}
         </p>
       )}
@@ -54,9 +38,6 @@ export function GeoMapOverlays({
   description,
   tooltipClassName,
   popupClassName,
-  popupContentClassName,
-  popupTitleClassName,
-  popupDescriptionClassName,
 }: {
   tooltipMode: "none" | "hover" | "always";
   tooltipContent?: string;
@@ -64,9 +45,6 @@ export function GeoMapOverlays({
   description?: string;
   tooltipClassName?: string;
   popupClassName?: string;
-  popupContentClassName?: string;
-  popupTitleClassName?: string;
-  popupDescriptionClassName?: string;
 }) {
   const hasPopup = Boolean(label || description);
   const [isPopupOpen, setIsPopupOpen] = useState(false);
@@ -100,13 +78,7 @@ export function GeoMapOverlays({
           maxWidth={288}
           eventHandlers={popupEventHandlers}
         >
-          <GeoMapPopupContent
-            label={label}
-            description={description}
-            className={popupContentClassName}
-            titleClassName={popupTitleClassName}
-            descriptionClassName={popupDescriptionClassName}
-          />
+          <GeoMapPopupContent label={label} description={description} />
         </Popup>
       )}
     </>
