--- components/tool-ui/geo-map/geo-map-runtime.ts
@@ -1,4 +1,0 @@
-export type LeafletRuntime = Pick<
-  typeof import("leaflet"),
-  "divIcon" | "latLngBounds"
->;
