+++ public/r/registry.json
@@ -542,6 +542,11 @@
           "target": "components/tool-ui/geo-map/_adapter.tsx"
         },
         {
+          "path": "components/tool-ui/geo-map/geo-map-engine.tsx",
+          "type": "registry:component",
+          "target": "components/tool-ui/geo-map/geo-map-engine.tsx"
+        },
+        {
           "path": "components/tool-ui/geo-map/geo-map-icons.ts",
           "type": "registry:lib",
           "target": "components/tool-ui/geo-map/geo-map-icons.ts"
@@ -552,19 +557,9 @@
           "target": "components/tool-ui/geo-map/geo-map-overlays.tsx"
         },
         {
-          "path": "components/tool-ui/geo-map/geo-map-runtime.ts",
-          "type": "registry:lib",
-          "target": "components/tool-ui/geo-map/geo-map-runtime.ts"
-        },
-        {
           "path": "components/tool-ui/geo-map/geo-map-theme.module.css",
           "type": "registry:style",
           "target": "components/tool-ui/geo-map/geo-map-theme.module.css"
-        },
-        {
-          "path": "components/tool-ui/geo-map/geo-map-viewport.tsx",
-          "type": "registry:component",
-          "target": "components/tool-ui/geo-map/geo-map-viewport.tsx"
         },
         {
           "path": "components/tool-ui/geo-map/geo-map.tsx",
@@ -585,11 +580,6 @@
           "path": "components/tool-ui/geo-map/schema.ts",
           "type": "registry:lib",
           "target": "components/tool-ui/geo-map/schema.ts"
-        },
-        {
-          "path": "components/tool-ui/geo-map/spatial.ts",
-          "type": "registry:lib",
-          "target": "components/tool-ui/geo-map/spatial.ts"
         },
         {
           "path": "components/tool-ui/shared/contract.ts",
